refactor(apis): split sticker types to Sticker model
