+++ src/apis/getStickerCategoryDetail.ts
@@ -1,28 +1,9 @@
 import { ZaloApiError } from "../Errors/ZaloApiError.js";
 import { apiFactory } from "../utils.js";
 
-export type GetStickerCategoryDetailResponse = {
-    id: number;
-    cateId: number;
-    type: number;
-    text: string;
-    uri: string;
-    fkey: number;
-    status: number;
-    stickerUrl: string;
-    stickerSpriteUrl: string;
-    stickerWebpUrl: string | null;
-    totalFrames: number;
-    duration: number;
-    effectId: number;
-    checksum: string;
-    ext: number;
-    source: number;
-    fss: null;
-    fssInfo: null;
-    version: number;
-    extInfo: null;
-}[];
+import type { StickerDetail } from "../models/index.js";
+
+export type GetStickerCategoryDetailResponse = StickerDetail[];
 
 export const getStickerCategoryDetailFactory = apiFactory<GetStickerCategoryDetailResponse>()((api, ctx, utils) => {
     const serviceURL = utils.makeURL(`${api.zpwServiceMap.sticker[0]}/api/message/sticker/category/sticker_detail`);
