+++ src/models/Sticker.ts
@@ -1,0 +1,28 @@
+export type StickerDetail = {
+    id: number;
+    cateId: number;
+    type: number;
+    text: string;
+    uri: string;
+    fkey: number;
+    status: number;
+    stickerUrl: string;
+    stickerSpriteUrl: string;
+    stickerWebpUrl: string | null;
+    totalFrames: number;
+    duration: number;
+    effectId: number;
+    checksum: string;
+    ext: number;
+    source: number;
+    fss: unknown;
+    fssInfo: unknown;
+    version: number;
+    extInfo: unknown;
+};
+
+export type StickerBasic = {
+    type: number;
+    cate_id: number;
+    sticker_id: number;
+};
