+++ src/models/index.ts
@@ -18,3 +18,4 @@ export * from "./Undo.js";
 export * from "./User.js";
 export * from "./ZBusiness.js";
 export * from "./Label.js";
+export * from "./Sticker.js";
